backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# API & File Handling
python-multipart==0.0.6
aiofiles==23.2.1
python-magic==0.4.27

# Google Cloud
google-cloud-vision==3.5.0
google-auth==2.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1